Remove Qwen3VL_PositionIDs, Qwen3VLConfiguration, Qwen3VLLanguage, and Qwen3VLVision files
