--- Libraries/MLXVLM/Models/Qwen3VLConfiguration.swift
@@ -1,168 +1,0 @@
-// Copyright © 2025 Apple Inc.
-
-import Foundation
-
-public struct Qwen3VLConfiguration: Codable, Sendable {
-
-    public struct TextConfiguration: Codable, Sendable {
-        public let modelType: String
-        public let hiddenSize: Int
-        public let intermediateSize: Int
-        public let numHiddenLayers: Int
-        public let numAttentionHeads: Int
-        private let _numKeyValueHeads: Int?
-        public var numKeyValueHeads: Int { _numKeyValueHeads ?? numAttentionHeads }
-        public let headDim: Int
-        private let _ropeTheta: Double?
-        public var ropeTheta: Double { _ropeTheta ?? 1_000_000 }
-        public let maxPositionEmbeddings: Int
-        private let _rmsNormEps: Double?
-        public var rmsNormEps: Double { _rmsNormEps ?? 1e-6 }
-        private let _ropeScaling: RoPEScaling?
-        public var ropeScaling: RoPEScaling? { _ropeScaling }
-        private let _normTopKProb: Bool?
-        public var normTopKProb: Bool { _normTopKProb ?? true }
-        private let _tieWordEmbeddings: Bool?
-        public var tieWordEmbeddings: Bool { _tieWordEmbeddings ?? true }
-        private let _attentionBias: Bool?
-        public var attentionBias: Bool { _attentionBias ?? false }
-        private let _hiddenAct: String?
-        public var hiddenAct: String { _hiddenAct ?? "silu" }
-        public let vocabSize: Int
-
-        enum CodingKeys: String, CodingKey {
-            case modelType = "model_type"
-            case hiddenSize = "hidden_size"
-            case intermediateSize = "intermediate_size"
-            case numHiddenLayers = "num_hidden_layers"
-            case numAttentionHeads = "num_attention_heads"
-            case _numKeyValueHeads = "num_key_value_heads"
-            case headDim = "head_dim"
-            case _ropeTheta = "rope_theta"
-            case maxPositionEmbeddings = "max_position_embeddings"
-            case _rmsNormEps = "rms_norm_eps"
-            case _ropeScaling = "rope_scaling"
-            case _normTopKProb = "norm_topk_prob"
-            case _tieWordEmbeddings = "tie_word_embeddings"
-            case _attentionBias = "attention_bias"
-            case _hiddenAct = "hidden_act"
-            case vocabSize = "vocab_size"
-        }
-    }
-
-    public struct VisionConfiguration: Codable, Sendable {
-        public let modelType: String
-        public let depth: Int
-        public let hiddenSize: Int
-        public let intermediateSize: Int
-        public let outHiddenSize: Int
-        public let numHeads: Int
-        public let patchSize: Int
-        public let spatialMergeSize: Int
-        public let temporalPatchSize: Int
-        public let numPositionEmbeddings: Int
-        private let _inChannels: Int?
-        public var inChannels: Int { _inChannels ?? 3 }
-        private let _hiddenAct: String?
-        public var hiddenAct: String { _hiddenAct ?? "gelu" }
-        private let _deepstackVisualIndexes: [Int]?
-        public var deepstackVisualIndexes: [Int] { _deepstackVisualIndexes ?? [] }
-
-        enum CodingKeys: String, CodingKey {
-            case modelType = "model_type"
-            case depth
-            case hiddenSize = "hidden_size"
-            case intermediateSize = "intermediate_size"
-            case outHiddenSize = "out_hidden_size"
-            case numHeads = "num_heads"
-            case patchSize = "patch_size"
-            case spatialMergeSize = "spatial_merge_size"
-            case temporalPatchSize = "temporal_patch_size"
-            case numPositionEmbeddings = "num_position_embeddings"
-            case _inChannels = "in_channels"
-            case _hiddenAct = "hidden_act"
-            case _deepstackVisualIndexes = "deepstack_visual_indexes"
-        }
-    }
-
-    public struct RoPEScaling: Codable, Sendable {
-        public let type: String?
-        public let mropeInterleaved: Bool?
-        public let mropeSection: [Int]?
-
-        enum CodingKeys: String, CodingKey {
-            case type
-            case mropeInterleaved = "mrope_interleaved"
-            case mropeSection = "mrope_section"
-        }
-
-        public init(type: String? = nil, mropeInterleaved: Bool? = nil, mropeSection: [Int]? = nil)
-        {
-            self.type = type
-            self.mropeInterleaved = mropeInterleaved
-            self.mropeSection = mropeSection
-        }
-    }
-
-    public let textConfiguration: TextConfiguration
-    public let visionConfiguration: VisionConfiguration
-    public let modelType: String
-    private let _ignoreIndex: Int?
-    public var ignoreIndex: Int { _ignoreIndex ?? -100 }
-    private let _imageTokenId: Int?
-    public var imageTokenId: Int { _imageTokenId ?? 151_655 }
-    private let _videoTokenId: Int?
-    public var videoTokenId: Int { _videoTokenId ?? 151_656 }
-    private let _imageTokenIndex: Int?
-    public var imageTokenIndex: Int { _imageTokenIndex ?? imageTokenId }
-    private let _videoTokenIndex: Int?
-    public var videoTokenIndex: Int { _videoTokenIndex ?? videoTokenId }
-    private let _visionStartTokenId: Int?
-    public var visionStartTokenId: Int { _visionStartTokenId ?? 151_652 }
-    private let _visionEndTokenId: Int?
-    public var visionEndTokenId: Int { _visionEndTokenId ?? 151_653 }
-    private let _visionTokenId: Int?
-    public var visionTokenId: Int { _visionTokenId ?? 151_654 }
-    private let _vocabSize: Int?
-    public var vocabSize: Int { _vocabSize ?? textConfiguration.vocabSize }
-    private let _eosTokenId: [Int]?
-    public var eosTokenId: [Int]? { _eosTokenId }
-
-    enum CodingKeys: String, CodingKey {
-        case textConfiguration = "text_config"
-        case visionConfiguration = "vision_config"
-        case modelType = "model_type"
-        case _ignoreIndex = "ignore_index"
-        case _imageTokenId = "image_token_id"
-        case _videoTokenId = "video_token_id"
-        case _imageTokenIndex = "image_token_index"
-        case _videoTokenIndex = "video_token_index"
-        case _visionStartTokenId = "vision_start_token_id"
-        case _visionEndTokenId = "vision_end_token_id"
-        case _visionTokenId = "vision_token_id"
-        case _vocabSize = "vocab_size"
-        case _eosTokenId = "eos_token_id"
-    }
-
-    public init(
-        textConfiguration: TextConfiguration, visionConfiguration: VisionConfiguration,
-        modelType: String = "qwen3_vl", ignoreIndex: Int = -100, imageTokenId: Int = 151_655,
-        videoTokenId: Int = 151_656, imageTokenIndex: Int? = nil, videoTokenIndex: Int? = nil,
-        visionStartTokenId: Int = 151_652, visionEndTokenId: Int = 151_653,
-        visionTokenId: Int = 151_654, vocabSize: Int? = nil, eosTokenId: [Int]? = nil
-    ) {
-        self.textConfiguration = textConfiguration
-        self.visionConfiguration = visionConfiguration
-        self.modelType = modelType
-        self._ignoreIndex = ignoreIndex
-        self._imageTokenId = imageTokenId
-        self._videoTokenId = videoTokenId
-        self._imageTokenIndex = imageTokenIndex
-        self._videoTokenIndex = videoTokenIndex
-        self._visionStartTokenId = visionStartTokenId
-        self._visionEndTokenId = visionEndTokenId
-        self._visionTokenId = visionTokenId
-        self._vocabSize = vocabSize
-        self._eosTokenId = eosTokenId
-    }
-}
