--- Libraries/MLXVLM/Models/Qwen3VL_PositionIDs.swift
@@ -1,192 +1,0 @@
-// Copyright © 2025 Apple Inc.
-
-import Foundation
-import MLX
-import MLXLMCommon
-
-extension Qwen3VLLanguage {
-
-    static func getRopeIndex(
-        inputIds: MLXArray,
-        imageGridTHW: [THW]?,
-        videoGridTHW: [THW]?,
-        spatialMergeSize: Int,
-        imageTokenId: Int,
-        videoTokenId: Int,
-        visionStartTokenId: Int,
-        attentionMask: MLXArray? = nil
-    ) -> (MLXArray, MLXArray) {
-
-        let (batchSize, seqLength) = (inputIds.dim(0), inputIds.dim(1))
-
-        var positionIds = MLXArray(0 ..< seqLength).asType(.int32)
-        positionIds = broadcast(positionIds[.newAxis, 0...], to: [batchSize, seqLength])
-
-        guard inputIds.ndim > 0, imageGridTHW != nil || videoGridTHW != nil else {
-            let positionIds3D = broadcast(
-                positionIds[.newAxis, 0..., 0...], to: [3, batchSize, seqLength])
-            let zeros = MLXArray.zeros([batchSize], dtype: .int32)
-            return (positionIds3D, zeros)
-        }
-
-        positionIds = ones(like: inputIds).asType(.int32)
-        positionIds = broadcast(positionIds[.newAxis, 0..., 0...], to: [3, batchSize, seqLength])
-
-        var mropePositionDeltas: [Int] = []
-        let mask = attentionMask ?? ones(like: inputIds)
-
-        // Process each batch item (assume batch=1 for now)
-        for batchIdx in 0 ..< batchSize {
-            var batchInputIds = inputIds[batchIdx, 0...]
-
-            // Mask out padding - use where from MLX module
-            batchInputIds = `where`(
-                mask[batchIdx, 0...] .== 1, batchInputIds, zeros(like: batchInputIds))
-
-            // Count images and videos in this sequence
-            let visionStartMask = (batchInputIds .== MLXArray(visionStartTokenId))
-            let visionStartWeighted = `where`(
-                visionStartMask, MLXArray(0 ..< seqLength), zeros(like: batchInputIds))
-            let visionStartIdx = argMax(visionStartWeighted).item(Int.self)
-
-            guard visionStartIdx < seqLength - 1 else {
-                continue  // No vision tokens
-            }
-
-            let imageNums = ((batchInputIds .== MLXArray(imageTokenId)).asType(.int32).sum()).item(
-                Int.self)
-            let videoNums = ((batchInputIds .== MLXArray(videoTokenId)).asType(.int32).sum()).item(
-                Int.self)
-
-            let inputTokens = batchInputIds.asArray(Int32.self).map { Int($0) }
-            var llmPosIdsList: [MLXArray] = []
-
-            var st = 0
-            var remainImages = imageNums
-            var remainVideos = videoNums
-            var imageIndex = 0
-            var videoIndex = 0
-
-            // Process each image/video in sequence
-            for _ in 0 ..< (imageNums + videoNums) {
-                // Find next image/video token position
-                let edImage: Int
-                if remainImages > 0, let idx = inputTokens[st...].firstIndex(of: imageTokenId) {
-                    edImage = idx
-                } else {
-                    edImage = inputTokens.count + 1
-                }
-
-                let edVideo: Int
-                if remainVideos > 0, let idx = inputTokens[st...].firstIndex(of: videoTokenId) {
-                    edVideo = idx
-                } else {
-                    edVideo = inputTokens.count + 1
-                }
-
-                let (t, h, w, ed): (Int, Int, Int, Int)
-                if edImage < edVideo {
-                    // Process image
-                    guard let grid = imageGridTHW, imageIndex < grid.count else { break }
-                    (t, h, w) = grid[imageIndex].values
-                    imageIndex += 1
-                    remainImages -= 1
-                    ed = edImage
-                } else {
-                    // Process video
-                    guard let grid = videoGridTHW, videoIndex < grid.count else { break }
-                    (t, h, w) = grid[videoIndex].values
-                    videoIndex += 1
-                    remainVideos -= 1
-                    ed = edVideo
-                }
-
-                let llmGridT = t
-                let llmGridH = h / spatialMergeSize
-                let llmGridW = w / spatialMergeSize
-
-                // Calculate starting index
-                let stIdx: Int
-                if let lastArray = llmPosIdsList.last {
-                    let maxVal = lastArray.max().item(Int.self)
-                    stIdx = maxVal + 1
-                } else {
-                    stIdx = 0
-                }
-
-                // Add text tokens before this visual block
-                let textLen = ed - st
-                if textLen > 0 {
-                    var index = MLXArray(0 ..< textLen).reshaped([1, textLen])
-                    index = broadcast(index, to: [3, textLen])
-                    index = index + MLXArray(stIdx)
-                    llmPosIdsList.append(index)
-                }
-
-                // Add 3D position IDs for visual tokens
-                // Python: mx.stack([t_index, h_index, w_index]) + text_len + st_idx
-                // Adds offset to ALL three dimensions!
-                var tIndex = MLXArray(0 ..< llmGridT).reshaped([llmGridT, 1])
-                tIndex = broadcast(tIndex, to: [llmGridT, llmGridH * llmGridW])
-                tIndex = tIndex.flattened()
-
-                var hIndex = MLXArray(0 ..< llmGridH).reshaped([1, llmGridH, 1])
-                hIndex = broadcast(hIndex, to: [llmGridT, llmGridH, llmGridW])
-                hIndex = hIndex.flattened()
-
-                var wIndex = MLXArray(0 ..< llmGridW).reshaped([1, 1, llmGridW])
-                wIndex = broadcast(wIndex, to: [llmGridT, llmGridH, llmGridW])
-                wIndex = wIndex.flattened()
-
-                let visualPosIds = stacked([tIndex, hIndex, wIndex]) + MLXArray(textLen + stIdx)
-                llmPosIdsList.append(visualPosIds)
-
-                st = ed + llmGridT * llmGridH * llmGridW
-            }
-
-            // Add remaining text tokens after last visual block
-            if st < inputTokens.count {
-                let stIdx: Int
-                if let lastArray = llmPosIdsList.last {
-                    let maxVal = lastArray.max().item(Int.self)
-                    stIdx = maxVal + 1
-                } else {
-                    stIdx = 0
-                }
-
-                let textLen = inputTokens.count - st
-                var tIndex = MLXArray(0 ..< textLen).reshaped([1, textLen])
-                tIndex = broadcast(tIndex, to: [3, textLen])
-                llmPosIdsList.append(tIndex + MLXArray(stIdx))
-            }
-
-            // Concatenate all position IDs for this batch item
-            if !llmPosIdsList.isEmpty {
-                let llmPositions = concatenated(llmPosIdsList, axis: 1)  // [3, seq]
-
-                // Update position_ids for this batch
-                let expandedMask = broadcast(
-                    mask[batchIdx, 0...][.newAxis, .newAxis, 0...], to: [3, 1, seqLength])
-                let expandedPositions = llmPositions[0..., .newAxis, 0...]
-                let newPositions = `where`(
-                    expandedMask, expandedPositions,
-                    positionIds[0..., batchIdx ..< batchIdx + 1, 0...])
-
-                // Replace this batch's position IDs (assumes batch size = 1)
-                positionIds = newPositions
-
-                let maxPosId = llmPositions.max().item(Int.self)
-                mropePositionDeltas.append(maxPosId + 1 - inputTokens.count)
-            }
-        }
-
-        // Python always returns deltas array (zeros for text-only, computed values for multimodal)
-        let deltas: MLXArray
-        if mropePositionDeltas.isEmpty {
-            deltas = MLXArray.zeros([batchSize], dtype: .int32)
-        } else {
-            deltas = MLXArray(mropePositionDeltas.map { Int32($0) })
-        }
-        return (positionIds, deltas)
-    }
-}
